% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_old.R
\name{AllowFileManipulationFromInitialiseProject}
\alias{AllowFileManipulationFromInitialiseProject}
\title{Allows for InitialiseProject to create folders and
delete empty folders on your computer (depreciated)}
\usage{
AllowFileManipulationFromInitialiseProject()
}
\description{
Allows for InitialiseProject to create folders and
delete empty folders on your computer (depreciated)
}
