% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{cat_to_filepath_function_factory}
\alias{cat_to_filepath_function_factory}
\title{Create a function to write to a specific file}
\usage{
cat_to_filepath_function_factory(filepath)
}
\arguments{
\item{filepath}{The path to the file to write to.}
}
\value{
A function that writes to the specified file. It takes these
parameters:
\itemize{
\item \code{...}: Content to write.
\item \code{sep}: Separator between elements (default: "").
\item \code{append}: Whether to append to existing content (default: TRUE).
}
}
\description{
Creates a closure that writes to a specified file path. Use it to create
several functions that write to different files with the same behavior.
}
\keyword{internal}
