% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc.R
\name{check_max_loc_per_file}
\alias{check_max_loc_per_file}
\title{Error if any file holds more code lines than the limit}
\usage{
check_max_loc_per_file(paths, max_loc)
}
\arguments{
\item{paths}{Character vector of folder paths to check.}

\item{max_loc}{The maximum number of code lines a single file may hold.}
}
\description{
Error if any file holds more code lines than the limit
}
\keyword{internal}
