% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_installed.R
\name{check_missing_packages}
\alias{check_missing_packages}
\title{Error if packages are missing}
\usage{
check_missing_packages(paths)
}
\arguments{
\item{paths}{Character vector of file or directory paths to scan.}
}
\description{
Error if packages are missing
}
\keyword{internal}
