% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project_quarto}
\alias{create_project_quarto}
\title{Create example project that uses quarto}
\usage{
create_project_quarto(home, results)
}
\arguments{
\item{home}{Location of the 'home' directory.}

\item{results}{Location of the 'results' directory.}
}
\description{
Create example project that uses quarto
}
