% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project_quarto_external_results}
\alias{create_project_quarto_external_results}
\title{Create example project that uses quarto with results generated outside the .qmd file}
\usage{
create_project_quarto_external_results(home, results)
}
\arguments{
\item{home}{Location of the 'home' directory.}

\item{results}{Location of the 'results' directory.}
}
\description{
Create example project that uses quarto with results generated outside the .qmd file
}
