% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_installed.R
\name{find_missing_packages}
\alias{find_missing_packages}
\title{Find missing packages referenced in R files}
\usage{
find_missing_packages(paths)
}
\arguments{
\item{paths}{Character vector of file or directory paths to scan.}
}
\value{
Character vector of missing package names.
}
\description{
Find missing packages referenced in R files
}
\keyword{internal}
