% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_project.R
\name{initialize_project_folders}
\alias{initialize_project_folders}
\title{Initialize project folder structure}
\usage{
initialize_project_folders(
  env,
  home,
  results,
  encode_from,
  encode_to,
  proj,
  ...
)
}
\arguments{
\item{env}{Environment to source code into.}

\item{home}{Home directory path.}

\item{results}{Results directory path.}

\item{encode_from}{Source encoding.}

\item{encode_to}{Target encoding.}

\item{proj}{Project environment.}

\item{...}{Additional folder arguments.}
}
\description{
Initialize project folder structure
}
\keyword{internal}
