% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc.R
\name{loc_one_file}
\alias{loc_one_file}
\title{Count the code lines in one R file}
\usage{
loc_one_file(path)
}
\arguments{
\item{path}{Path to a single R file.}
}
\value{
An integer.
}
\description{
Count the code lines in one R file
}
\keyword{internal}
