% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{move_file_or_dir}
\alias{move_file_or_dir}
\title{Move one file or directory}
\usage{
move_file_or_dir(from, to)
}
\arguments{
\item{from}{Filepath or directory path.}

\item{to}{Filepath or directory path.}
}
\description{
Move one file or directory
}
