% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_old.R, R/initialize_project.R
\name{AllowFileManipulationFromInitialiseProject}
\alias{AllowFileManipulationFromInitialiseProject}
\title{Allows for InitialiseProject to create folders and
delete empty folders on your computer}
\usage{
AllowFileManipulationFromInitialiseProject()

AllowFileManipulationFromInitialiseProject()
}
\description{
Allows for InitialiseProject to create folders and
delete empty folders on your computer

Allows for InitialiseProject to create folders and
delete empty folders on your computer
}
