% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{ls_files}
\alias{ls_files}
\title{List files and directories
Equivalent to the unix \code{ls} command.}
\usage{
ls_files(path = ".", regexp = NULL)
}
\arguments{
\item{path}{A character vector of one or more paths.}

\item{regexp}{A regular expression that is passed to \code{list.files}.}
}
\value{
filepaths and directory paths as a character vector
}
\description{
List files and directories
Equivalent to the unix \code{ls} command.
}
