% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{move_directory}
\alias{move_directory}
\title{Move directory}
\usage{
move_directory(from, to, overwrite_to = FALSE)
}
\arguments{
\item{from}{Filepath or directory path.}

\item{to}{Filepath or directory path.}

\item{overwrite_to}{Boolean.}
}
\description{
Move directory
}
