% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_installed.R
\name{package_installed}
\alias{package_installed}
\title{Is a package installed?}
\usage{
package_installed(pkg)
}
\arguments{
\item{pkg}{Package name.}
}
\value{
Boolean. True if package is installed, false if not.
}
\description{
Is a package installed?
}
\examples{
org::package_installed("data.table")
}
