% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{path}
\alias{path}
\title{Construct path to a file or directory}
\usage{
path(...)
}
\arguments{
\item{...}{Character vectors that will be concatenated with "/" as a separator.}
}
\description{
Construct path to a file or directory
}
