% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_env.R
\docType{data}
\name{project}
\alias{project}
\title{Project folder locations}
\format{
An environment containing the following elements:
\describe{
\item{home}{The folder containing 'Run.R' and 'R/'}
\item{results_today}{The folder inside \code{results} with today's date, created by \code{initialize_project}}
}
}
\usage{
project
}
\description{
An environment that stores the locations of folders used in the project.
}
\keyword{datasets}
