% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_env.R
\docType{data}
\name{project}
\alias{project}
\title{Folders to be used/referenced (environment)}
\format{
An object of class \code{environment} of length 0.
}
\usage{
project
}
\description{
Folders to be used/referenced (environment)
}
\keyword{datasets}
