% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_project.R
\name{select_folder_that_exists}
\alias{select_folder_that_exists}
\title{Select first existing folder from a list}
\usage{
select_folder_that_exists(folders, name)
}
\arguments{
\item{folders}{Character vector of folder paths to check}

\item{name}{Name of the folder type for error messages}
}
\value{
List with folder path and id
}
\description{
Select first existing folder from a list
}
\keyword{internal}
