% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_project.R
\name{set_results}
\alias{set_results}
\title{Set results folder after initialization}
\usage{
set_results(results)
}
\arguments{
\item{results}{A folder inside `results` with today's date will be created and it will be accessible via `org::project$results_today` (this is where you will store all of your results)}
}
\description{
Set results folder after initialization
}
