% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_project.R
\name{set_results}
\alias{set_results}
\title{Set results folder after initialization}
\usage{
set_results(results, proj = org::project)
}
\arguments{
\item{results}{A character vector specifying the results folder path(s).}

\item{proj}{The project environment. Default is \code{org::project}.}
}
\value{
Nothing. Alters the \code{proj} environment to include \verb{$results} and \verb{$results_today}.
}
\description{
This function sets the results folder in the project environment.
A folder with today's date will be created inside the results folder.
}
