% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_project.R
\name{set_results_internal}
\alias{set_results_internal}
\title{Internal function to set results folder}
\usage{
set_results_internal(results, proj)
}
\arguments{
\item{results}{Results folder path}

\item{proj}{Project environment}
}
\description{
Internal function to set results folder
}
\keyword{internal}
