% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_project.R
\name{source_to_environment}
\alias{source_to_environment}
\title{Source R files into environment}
\usage{
source_to_environment(
  proj,
  env,
  folders_to_be_sourced,
  source_folders_absolute
)
}
\arguments{
\item{proj}{Project environment}

\item{env}{Target environment for sourcing}

\item{folders_to_be_sourced}{Folders containing R files}

\item{source_folders_absolute}{Whether folder paths are absolute}
}
\description{
Source R files into environment
}
\keyword{internal}
