% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_text.R
\name{write_text}
\alias{write_text}
\title{Write text to a file}
\usage{
write_text(
  txt,
  file = "",
 
    header = "**THIS FILE IS CONSTANTLY OVERWRITTEN -- DO NOT MANUALLY EDIT**\\r\\n\\r\\n"
)
}
\arguments{
\item{txt}{Text to be written}

\item{file}{File, passed through to \code{base::cat}}

\item{header}{Optional header that is inserted at the top of the text file}
}
\value{
No return value.
}
\description{
Write text to a file
}
